[refactor]: organizando archivos en subcarpetas y corrigiendo imports
