+++ frontend/src/components/NavBar.js
@@ -2,7 +2,7 @@
 //REACT HOOKS/IMPORTS
 import { Link } from "react-router-dom";
 //CUSTOM HOOKS
-import { useLogout } from "../hooks/useLogout";
+import { useLogout } from "../hooks/usuarioHooks/useLogout";
 import { useAuthContext } from "../hooks/useAuthContext";
 //COMPONENTS
 import carwashlogo from "../images/carwash-logo.jpg";
