+++ frontend/src/hooks/usuarioHooks/useLogout.js
@@ -1,6 +1,6 @@
 //************************** IMPORTED
 //CUSTOM HOOKS
-import { useAuthContext } from "../hooks/useAuthContext";
+import { useAuthContext } from "../useAuthContext";
 //ENV VARIBLES
 const apiURL = process.env.REACT_APP_DEVURL;
 //**************************************************************
